edv::cli

Struct CommandRegistry

pub struct CommandRegistry {
    commands: HashMap<String, Box<dyn Command>>,
}

Registry for managing available commands.

The command registry maintains a collection of all available commands and provides methods for registering, looking up, and listing commands.

Fields
commands: HashMap<String, Box<dyn Command>>

Map of command names to command implementations

Implementations

impl CommandRegistry

pub fn new() -> Self

Creates a new, empty command registry.

Returns

A new CommandRegistry instance.

pub fn register(&mut self, command: Box<dyn Command>) -> Result<()>

Registers a command with the registry.

Arguments

command - The command to register

Returns

Result<()> indicating success or whether the command name was already registered.

Errors

Returns an error if a command with the same name is already registered.

pub fn get(&self, name: &str) -> Result<&dyn Command>

Gets a command by name.

Arguments

name - The name of the command to retrieve

Returns

Result<&dyn Command> containing a reference to the command if found.

Errors

Returns an error if no command with the given name is registered.

pub fn list(&self) -> Vec<&dyn Command>

Gets a list of all registered commands.

Returns

A vector of references to all registered commands.

pub fn command_names(&self) -> Vec<String>

Gets a list of all command names.

Returns

A vector of command names.

pub fn has_command(&self, name: &str) -> bool

Checks if a command with the given name is registered.

Arguments

name - The name of the command to check

Returns

true if a command with the given name is registered, false otherwise.

pub fn command_count(&self) -> usize

Gets the number of registered commands.

Returns

The number of commands in the registry.