edv::project::timeline
Struct Timeline

pub struct Timeline { /* private fields */ }

A timeline containing multiple tracks.

Implementations

impl Timeline

pub fn new() -> Self

Creates a new empty timeline.

pub fn get_tracks(&self) -> &[Track]

Gets a reference to all tracks in the timeline.

pub fn get_track_mut(&mut self, track_id: TrackId) -> Option<&mut Track>

Gets a mutable reference to a track by its ID.

pub fn get_track(&self, track_id: TrackId) -> Option<&Track>

Gets a reference to a track by its ID.

pub fn has_track(&self, track_id: TrackId) -> bool

Checks if a track with the given ID exists.

pub fn find_track_containing_clip(&self, clip_id: ClipId) -> Option<TrackId>

Finds the ID of the track containing the specified clip.

Arguments

clip_id - The ID of the clip to search for.

Returns

An Option<TrackId> containing the ID of the track if found, otherwise None.

pub fn add_track(&mut self, kind: TrackKind) -> TrackId

Adds a new track to the timeline.

Arguments

kind - The kind of track to add

Returns

The ID of the added track.

pub fn remove_track(&mut self, track_id: TrackId) -> Result<()>

Removes a track from the timeline.

Arguments

track_id - The ID of the track to remove

Returns

Ok(()) if the track was removed successfully, or an error if the track does not exist or has dependencies.

Errors

Returns an error if the track does not exist or if there are other tracks that depend on it.

pub fn add_clip(&mut self, track_id: TrackId, clip: Clip) -> Result<()>

Adds a clip to a track.

Arguments

track_id - The ID of the track to add the clip to

clip - The clip to add

Returns

Ok(()) if the clip was added successfully, or an error if the track does not exist or the clip overlaps with an existing clip.

Errors

Returns an error if:

The track does not exist

The clip overlaps with an existing clip in the track

pub fn remove_clip(&mut self, track_id: TrackId, clip_id: ClipId) -> Result<()>

Removes a clip from a track.

Arguments

track_id - The ID of the track to remove the clip from

clip_id - The ID of the clip to remove

Returns

Ok(()) if the clip was removed successfully, or an error if the track or clip does not exist.

Errors

Returns an error if:

The track does not exist

The clip does not exist in the track

pub fn multi_track_manager(&self) -> &MultiTrackManager

Gets the multi-track manager.

pub fn multi_track_manager_mut(&mut self) -> &mut MultiTrackManager

Gets a mutable reference to the multi-track manager.

pub fn duration(&self) -> Duration

Gets the duration of the timeline.

pub fn split_clip( &mut self, track_id: TrackId, clip_id: ClipId, position: TimePosition, ) -> Result<ClipId>

Splits a clip at the specified position.

Arguments

track_id - The ID of the track containing the clip

clip_id - The ID of the clip to split

position - The position at which to split the clip

Returns

The ID of the new clip created from the split, or an error if the operation failed.

Errors

Returns an error if:

The track or clip is not found

The position is outside the clip

The clip cannot be split for some other reason

pub fn merge_clips( &mut self, track_id: TrackId, first_clip_id: ClipId, second_clip_id: ClipId, ) -> Result<()>

Merges two adjacent clips.

Arguments

track_id - The ID of the track containing the clips

first_clip_id - The ID of the first clip

second_clip_id - The ID of the second clip

Returns

Ok(()) if the clips were merged successfully, or an error if the operation failed.

Errors

Returns an error if:

The track or either clip is not found

The clips are not adjacent

The clips use different assets

The clips cannot be merged for some other reason

pub fn move_clip_to_track( &mut self, source_track_id: TrackId, target_track_id: TrackId, clip_id: ClipId, new_position: Option<TimePosition>, ) -> Result<()>

Moves a clip from one track to another.

Arguments

source_track_id - The ID of the track containing the clip

target_track_id - The ID of the track to move the clip to

clip_id - The ID of the clip to move

new_position - Optional new position for the clip (if None, keeps the original position)

Returns

Ok(()) if the clip was moved successfully, or an error if the operation failed.

Errors

Returns an error if:

The source track, target track, or clip is not found

The target track is of a different kind than the source track

Moving the clip would cause an overlap in the target track

The clip cannot be moved for some other reason