edv::project::timeline

Struct Timeline

pub struct Timeline {
    tracks: Vec<Track>,
    track_index_map: HashMap<TrackId, usize>,
    multi_track_manager: MultiTrackManager,
    history: EditHistory,
}

Timeline data structure.

Fields
tracks: Vec<Track>

Tracks in the timeline.

track_index_map: HashMap<TrackId, usize>

Mapping from track ID to index in the tracks vector.

multi_track_manager: MultiTrackManager

Multi-track relationship manager.

history: EditHistory

Edit history.

Implementations

impl Timeline

pub fn new() -> Self

Creates a new empty timeline.

pub fn get_tracks(&self) -> &[Track]

Gets a reference to all tracks in the timeline.

pub fn get_track_mut(&mut self, track_id: TrackId) -> Option<&mut Track>

Gets a mutable reference to a track by its ID.

pub fn get_track(&self, track_id: TrackId) -> Option<&Track>

Gets a reference to a track by its ID.

pub fn has_track(&self, track_id: TrackId) -> bool

Checks if a track with the given ID exists.

pub fn find_track_containing_clip(&self, clip_id: ClipId) -> Option<TrackId>

Finds the ID of the track containing the specified clip.

Arguments

clip_id - The ID of the clip to search for.

Returns

An Option<TrackId> containing the ID of the track if found, otherwise None.

pub fn add_track(&mut self, kind: TrackKind) -> TrackId

Adds a new track to the timeline.

Arguments

kind - The kind of track to add

Returns

The ID of the added track.

pub fn remove_track(&mut self, track_id: TrackId) -> Result<()>

Removes a track from the timeline.

Arguments

track_id - The ID of the track to remove

Returns

Ok(()) if the track was removed successfully, or an error if the track does not exist or has dependencies.

Errors

Returns an error if the track does not exist or if there are other tracks that depend on it.

pub fn add_clip(&mut self, track_id: TrackId, clip: Clip) -> Result<()>

Adds a clip to a track.

Arguments

track_id - The ID of the track to add the clip to

clip - The clip to add

Returns

Ok(()) if the clip was added successfully, or an error if the track does not exist or the clip overlaps with an existing clip.

Errors

Returns an error if:

The track does not exist

The clip overlaps with an existing clip in the track

pub fn add_clip_with_history( &mut self, track_id: TrackId, clip: Clip, ) -> Result<()>

Adds a clip to a track and records the action in history.

Arguments

track_id - The ID of the track to add the clip to

clip - The clip to add

Returns

Ok(()) if the clip was added successfully, or an error if the track does not exist or the clip overlaps with an existing clip.

Errors

Returns an error if:

The track does not exist

The clip overlaps with an existing clip in the track

pub fn remove_clip_with_history( &mut self, track_id: TrackId, clip_id: ClipId, ) -> Result<()>

Removes a clip from a track and records the action in history.

Arguments

track_id - The ID of the track to remove the clip from

clip_id - The ID of the clip to remove

Returns

Ok(()) if the clip was removed successfully, or an error if the track or clip does not exist.

Errors

Returns an error if:

The track does not exist

The clip does not exist in the track

pub fn remove_clip(&mut self, track_id: TrackId, clip_id: ClipId) -> Result<()>

Removes a clip from a track.

Arguments

track_id - The ID of the track to remove the clip from

clip_id - The ID of the clip to remove

Returns

Ok(()) if the clip was removed successfully, or an error if the track or clip does not exist.

Errors

Returns an error if:

The track does not exist

The clip does not exist in the track

pub fn multi_track_manager(&self) -> &MultiTrackManager

Gets the multi-track manager.

pub fn multi_track_manager_mut(&mut self) -> &mut MultiTrackManager

Gets a mutable reference to the multi-track manager.

pub fn duration(&self) -> Duration

Gets the duration of the timeline.

pub fn split_clip( &mut self, track_id: TrackId, clip_id: ClipId, position: TimePosition, ) -> Result<ClipId>

Splits a clip at the specified position.

Arguments

track_id - The ID of the track containing the clip

clip_id - The ID of the clip to split

position - The position at which to split the clip

Returns

The ID of the new clip created from the split, or an error if the operation failed.

Errors

Returns an error if:

The track or clip is not found

The position is outside the clip

The clip cannot be split for some other reason

pub fn merge_clips( &mut self, track_id: TrackId, first_clip_id: ClipId, second_clip_id: ClipId, ) -> Result<()>

Merges two adjacent clips.

Arguments

track_id - The ID of the track containing the clips

first_clip_id - The ID of the first clip

second_clip_id - The ID of the second clip

Returns

Ok(()) if the clips were merged successfully, or an error if the operation failed.

Errors

Returns an error if:

The track or either clip is not found

The clips are not adjacent

The clips use different assets

The clips cannot be merged for some other reason

pub fn move_clip_to_track( &mut self, source_track_id: TrackId, target_track_id: TrackId, clip_id: ClipId, new_position: Option<TimePosition>, ) -> Result<()>

Moves a clip from one track to another.

Arguments

source_track_id - The ID of the track containing the clip

target_track_id - The ID of the track to move the clip to

clip_id - The ID of the clip to move

new_position - Optional new position for the clip (if None, keeps the original position)

Returns

Ok(()) if the clip was moved successfully, or an error if the operation failed.

Errors

Returns an error if:

The source track, target track, or clip is not found

The target track is of a different kind than the source track

Moving the clip would cause an overlap in the target track

The clip cannot be moved for some other reason

pub fn move_clip_to_track_with_history( &mut self, source_track_id: TrackId, target_track_id: TrackId, clip_id: ClipId, new_position: Option<TimePosition>, ) -> Result<()>

Moves a clip to a different track and records the action in history.

Arguments

source_track_id - The ID of the track containing the clip

target_track_id - The ID of the track to move the clip to

clip_id - The ID of the clip to move

new_position - The new position for the clip, or None to keep the current position

Returns

Ok(()) if the clip was moved successfully, or an error if the source track, target track, or clip does not exist, or if the new position would cause overlap.

Errors

Returns an error if:

The source track does not exist

The target track does not exist

The clip does not exist in the source track

The new position would cause the clip to overlap with existing clips

pub fn history(&self) -> &EditHistory

Gets a reference to the edit history.

pub fn history_mut(&mut self) -> &mut EditHistory

Gets a mutable reference to the edit history.

pub fn undo(&mut self) -> Result<(), HistoryError>

Undoes the last edit action.

Returns

Ok(()) if the undo was successful, or an error if there is nothing to undo or if the undo operation failed.

Errors

Returns an error if:

There is nothing to undo

The undo operation failed

pub fn redo(&mut self) -> Result<(), HistoryError>

Redoes the last undone edit action.

Returns

Ok(()) if the redo was successful, or an error if there is nothing to redo or if the redo operation failed.

Errors

Returns an error if:

There is nothing to redo

The redo operation failed

pub fn can_undo(&self) -> bool

Checks if there are any actions that can be undone.

pub fn can_redo(&self) -> bool

Checks if there are any actions that can be redone.

pub fn begin_transaction( &mut self, description: Option<String>, ) -> Result<(), HistoryError>

Begins a transaction for grouping multiple edit actions.

Arguments

description - An optional description for the transaction

Returns

Ok(()) if the transaction was started successfully, or an error if a transaction is already in progress.

Errors

Returns an error if a transaction is already in progress.

pub fn commit_transaction(&mut self) -> Result<(), HistoryError>

Commits the current transaction.

Returns

Ok(()) if the transaction was committed successfully, or an error if no transaction is in progress.

Errors

Returns an error if no transaction is in progress.

pub fn rollback_transaction(&mut self) -> Result<(), HistoryError>

Rolls back the current transaction.

Returns

Ok(()) if the transaction was rolled back successfully, or an error if no transaction is in progress.

Errors

Returns an error if no transaction is in progress.

pub fn add_keyframe( &mut self, track_id: TrackId, property: &str, time: TimePosition, value: f64, easing: EasingFunction, ) -> Result<()>

Adds a keyframe to a track for a specific property.

Arguments

track_id - The ID of the track to add the keyframe to

property - The property name to animate

time - The time position of the keyframe

value - The value at the keyframe

easing - The easing function to use

Returns

A Result containing () if the keyframe was successfully added, or an error if the operation failed.

Errors

Returns an error if the track doesn’t exist or the keyframe couldn’t be added.

pub fn add_keyframe_with_history( &mut self, track_id: TrackId, property: &str, time: TimePosition, value: f64, easing: EasingFunction, ) -> Result<()>

Adds a keyframe to a track for a specific property and records the action in history.

Arguments

track_id - The ID of the track to add the keyframe to

property - The property name to animate

time - The time position of the keyframe

value - The value at the keyframe

easing - The easing function to use

Returns

A Result containing () if the keyframe was successfully added, or an error if the operation failed.

Errors

Returns an error if the track doesn’t exist or the keyframe couldn’t be added.

pub fn get_keyframe_value_at( &self, track_id: TrackId, property: &str, time: TimePosition, ) -> Option<f64>

Gets the animated property value at a specific time.

Arguments

track_id - The ID of the track

property - The property name

time - The time position to get the value at

Returns

An Option containing the interpolated value at the given time, or None if the track doesn’t exist, doesn’t have keyframe animation, or the property doesn’t have keyframes.

pub fn remove_keyframe( &mut self, track_id: TrackId, property: &str, time: TimePosition, ) -> Result<()>

Removes a keyframe from a track.

Arguments

track_id - The ID of the track

property - The property name

time - The time position of the keyframe to remove

Returns

A Result containing () if the keyframe was successfully removed, or an error if the operation failed.

Errors

Returns an error if the track doesn’t exist, the track doesn’t have keyframe animation, or the keyframe couldn’t be found.

pub fn remove_keyframe_with_history( &mut self, track_id: TrackId, property: &str, time: TimePosition, ) -> Result<()>

Removes a keyframe from a track and records the action in history.

Arguments

track_id - The ID of the track

property - The property name

time - The time position of the keyframe to remove

Returns

A Result containing () if the keyframe was successfully removed, or an error if the operation failed.

Errors

Returns an error if the track doesn’t exist, the track doesn’t have keyframe animation, or the keyframe couldn’t be found.

pub fn update_keyframe_with_history( &mut self, track_id: TrackId, property: &str, time: TimePosition, new_value: f64, new_easing: EasingFunction, ) -> Result<()>

Updates a keyframe on a track and records the action in history.

Arguments

track_id - The ID of the track

property - The property name

time - The time position of the keyframe

new_value - The new value for the keyframe

new_easing - The new easing function for the keyframe

Returns

A Result containing () if the keyframe was successfully updated, or an error if the operation failed.

Errors

Returns an error if the track doesn’t exist, the track doesn’t have keyframe animation, or the keyframe couldn’t be found.

pub fn validate(&self, assets: &[AssetReference]) -> Vec<TimelineError>

Validates the timeline for correctness and completeness.

This method performs a comprehensive check of the timeline to ensure it can be properly rendered:

Ensures there is at least one video track

Checks for overlapping clips in tracks

Validates all asset references

Checks track relationships

Returns

A list of validation errors, or an empty vector if the timeline is valid.

fn find_overlapping_clips(&self, track_id: TrackId) -> Vec<ClipId>

Finds clips that overlap in time within a track.

Arguments

track_id - The ID of the track to check

Returns

A vector of clip IDs that overlap with other clips.

fn has_gaps_between_clips(&self, track_id: TrackId) -> bool

Checks if there are gaps between clips in a track.

Arguments

track_id - The ID of the track to check

Returns

true if there are gaps between clips, false otherwise.